pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "image_container"
version = "0.0.1"
authors = [
  { name = "Ryuto Yoshida" },
]
description = "Unified Python interface for image containers with channel-order aware conversion."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "opencv-contrib-python",
    "pillow",
    "torch",
]

[project.urls]
Homepage = "https://github.com/ry-yoshida-private/ImageContainer"
"Bug Tracker" = "https://github.com/ry-yoshida-private/ImageContainer/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
container = ["py.typed"]
